Kept only ci_sanity tests for discovery tests
